feat: enabled CSV exporting for each project
